NetworkWebApplication/backend/requirements.txt
flask
flask-restful
pymongo
pythonping
apscheduler
python-dotenv
flask-cors
pytest
mongomock
requests
